restructure the file to be modular and in separate helper files/functions
